requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
ddddocr==1.5.6
pandas>=2.2.0
openpyxl>=3.1.2
xlrd>=2.0.1
Pillow>=10.0.0
python-docx>=1.1.0
PyYAML>=6.0.1
